Remove .py extension from fess_up tool
The fess_up tool had a .py extension which was causing import troubles
when it was installed. Tools run from /bin which tried to import the
fess_up module would actually import the script.

Removing the .py extension resolves this as Python will ignore non-.py
files when search for imports.

Resolves #12